max memory is 32-bit, for now

+++ src/wasm.h
@@ -932,7 +932,7 @@ public:
   size_t initial, max;
   std::vector<Segment> segments;
 
-  Memory() : initial(0), max(-1) {}
+  Memory() : initial(0), max((uint32_t)-1) {}
 };
 
 class Module {
